remove redundant class
